Bug fix.
Strange, some users were getting a NullReferenceException in RaiseExecuteChanged(). The only cause
I can think of is that the properties are changing _before_ the commands are initialized, but it's
strange, the viewmodel is created in the view constructor therefore no property should change.
This is the best fix I can think of.